refactor: move logic from screen to model, and use a future
