+++ lib/models/tabs/pdf_tab.dart
@@ -1,3 +1,5 @@
+import 'dart:isolate';
+
 import 'package:flutter/material.dart';
 import 'package:otzaria/models/books.dart';
 import 'package:otzaria/models/tabs/tab.dart';
@@ -42,6 +44,31 @@ class PdfBookTab extends OpenedTab {
       : super(book.title) {
     showLeftPane = ValueNotifier<bool>(openLeftPane);
     searchController.text = searchText;
+  }
+
+  Future<String?> currentSection() async {
+    if (outline.value == null) return null;
+
+    final currentPage = pdfViewerController.pageNumber;
+    if (currentPage == null) return null;
+
+    // Find the current section by looking through the outline
+    PdfOutlineNode? currentSection;
+
+    void searchOutline(List<PdfOutlineNode> entries, List<String> path) {
+      for (var entry in entries) {
+        if (entry.dest?.pageNumber != null &&
+            entry.dest!.pageNumber <= currentPage) {
+          currentSection = entry;
+          // Continue searching children to find most specific match
+          searchOutline(entry.children, [...path, entry.title]);
+        }
+      }
+    }
+
+    searchOutline(outline.value!, []);
+
+    return currentSection?.title;
   }
 
   /// Creates a new instance of [PdfBookTab] from a JSON map.
